fix compile error with IC14
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@953 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ RenderDevice.cpp
@@ -548,7 +548,7 @@ void RenderDevice::SetTextureFilter(DWORD texUnit, DWORD mode)
 		CHECKD3D(m_pD3DDevice->SetSamplerState(texUnit, D3DSAMP_MAGFILTER, m_mag_aniso ? D3DTEXF_ANISOTROPIC : D3DTEXF_LINEAR));
 		CHECKD3D(m_pD3DDevice->SetSamplerState(texUnit, D3DSAMP_MINFILTER, D3DTEXF_ANISOTROPIC));
 		CHECKD3D(m_pD3DDevice->SetSamplerState(texUnit, D3DSAMP_MIPFILTER, D3DTEXF_LINEAR));
-		CHECKD3D(m_pD3DDevice->SetSamplerState(texUnit, D3DSAMP_MAXANISOTROPY, min(m_maxaniso,16)));
+		CHECKD3D(m_pD3DDevice->SetSamplerState(texUnit, D3DSAMP_MAXANISOTROPY, min(m_maxaniso,(DWORD)16)));
 		break;
 	}
 }
